course_planner/requirements.txt
google-genai
streamlit
pypdf2
python-dotenv
openpyxl
